requirements.txt
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.26.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
fastapi==0.79.0
fastapi-pagination==0.9.3
h11==0.13.0
idna==3.3
iso8601==1.0.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pydantic==1.9.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-decouple==3.6
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
